+++ src/Separator.php
@@ -1,0 +1,20 @@
+<?php
+declare(strict_types=1);
+
+/**
+ * Text Table
+ *
+ * @author    Marcin Orlowski <mail (#) marcinOrlowski (.) com>
+ * @license   http://www.opensource.org/licenses/mit-license.php MIT
+ * @link      https://github.com/MarcinOrlowski/php-text-table
+ */
+
+namespace MarcinOrlowski\TextTable;
+
+/**
+ * Class represents single row of cells of the table.
+ */
+class Separator extends Row
+{
+    // Currently no logic in this class
+}